Revert "[direct-io-bulkfs-parquet]: shaded hadoop deps + bump parquet to 1.12.2"
This reverts commit a657d02d5c12313bab7ad8a403ba1c6d6b5afbdd.

+++ direct/io-bulkfs-parquet/src/main/java/cz/o2/proxima/direct/bulk/fs/parquet/ProximaParquetWriter.java
@@ -53,7 +53,7 @@ public class ProximaParquetWriter implements Writer {
             // For some reason Writer ignores settings this via withConf()
             .withCompressionCodec(compressionCodecName)
             .withRowGroupSize(
-                config.getLong(
+                config.getInt(
                     ParquetFileFormat.PARQUET_CONFIG_PAGE_SIZE_KEY_NAME,
                     ParquetFileFormat.PARQUET_DEFAULT_PAGE_SIZE))
             .withPageSize(
